requirements.txt
clickhouse-driver
polars
python-dotenv